packages/servicekit/pyproject.toml
[project]
name = "servicekit"
version = "0.1.0"
description = "Async SQLAlchemy database library for Python 3.13+ with FastAPI integration"
readme = "README.md"
authors = [{ name = "Morten Hansen", email = "morten@winterop.com" }]
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.13"
keywords = [
    "fastapi",
    "sqlalchemy",
    "async",
    "database",
    "rest-api",
    "crud",
    "vertical-slice",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
]
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.0",
    "fastapi[standard]>=0.119.0",
    "geojson-pydantic>=2.1.0",
    "gunicorn>=23.0.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-prometheus>=0.58b0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "pydantic>=2.12.0",
    "python-ulid>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "structlog>=24.4.0",
]

[project.urls]
Homepage = "https://github.com/winterop-com/servicekit"
Repository = "https://github.com/winterop-com/servicekit"
Issues = "https://github.com/winterop-com/servicekit/issues"
Documentation = "https://github.com/winterop-com/servicekit#readme"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
